Moved FailedLoginsModel to its own file
