+++ Project/Controllers/LoginController.cs
@@ -2,7 +2,6 @@
 using Project.LoginSystem;
 using Project.Models.LoginSystem;
 using System;
-using System.Diagnostics;
 using System.Linq;
 using System.Net.Http.Formatting;
 using System.Web.Http;
