--- Project/Models/LoginsModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace Project.Models
-{
-    public class LoginsModel
-    {
-        
-    }
-}
